to compile with gcc 4.3
git-svn-id: https://devel.codership.com/svn/galera/branches/dlopen@546 9148e567-b586-4176-b9b8-4e36cb4ec61b

+++ galera/src/dummy.c
@@ -56,7 +56,7 @@ static galera_status_t dg_disable(galera_t *g)
 static galera_status_t dg_recv(galera_t *g, void *ctx)
 {
     DBUG("");
-    sleep(LONG_MAX);
+    sleep(UINT_MAX);
     return GALERA_OK;
 }
